pyproject.toml
[project]
name = "chessticulate-api"
version = "0.2.0"
requires-python = ">=3.9"
dependencies = ["fastapi[all]", "sqlalchemy >= 2", "httpx", "python-dotenv", "py-bcrypt", "pyjwt", "asyncpg"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = ["black", "pylint", "pytest", "pytest-asyncio", "pytest-cov", "isort", "aiosqlite"]

[tool.isort]
line_length = 88

[tool.black]
line_length = 88
preview = true  # format long strings

[tool.pylint]
max-line-length = 88

[tool.pytest.ini_options]
addopts = "--cov=chessticulate_api"
